requirements.txt
boto3
pyyaml
requests
tabulate
